Extract the same behaviors as a command.
